Include symbols in package
Given that the new portable PDBs generated for NS projects are really small for the added benefit of 
useful stacktraces on errors, it would complement the sourcelink that's already in place. 

In particular, for VS extensions that use this package, including symbols in the package makes it easier 
for crash dumps to contain useful stacktraces. i.e. see https://github.com/jeffkl/ManagedDism/pull/76

This would fix bug http://work.azdo.io/971173 too :)